Small improvements to V4L2Camera and Timer
	modified:   io/V4L2Camera.cpp
	modified:   util/Timer.h
	modified:   ../tests/v4l2/v4l2.cpp

git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@228 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ cvt/util/Timer.h
@@ -32,7 +32,7 @@ namespace cvt {
 	    {
 		boost::posix_time::ptime now = boost::posix_time::microsec_clock::local_time();
 		boost::posix_time::time_duration t = now-time;
-		return t.total_microseconds();
+		return ( double ) t.total_microseconds();
 	    }
 
 	private:
